django/pyproject.toml
[tool.poetry]
name = "desafio-edesoft-basckend"
version = "0.1.0"
description = ""
authors = ["talesmm14 <talesmelquiades@hotmail.com>"]
readme = "README.md"
packages = [{include = "desafio_edesoft_basckend"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.7"
django-rest-framework = "^0.1.0"
django-environ = "^0.10.0"
boto3 = "^1.26.94"
awscli = "^1.27.94"
zappa = "^0.56.1"
pandas = "^1.5.3"
django-localflavor = "^3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
